+++ models_src/db_inits/tortoise_init.py
@@ -1,5 +1,9 @@
 from typing import Any
 
+import asyncpg
+from pgvector.asyncpg import register_vector
+from tortoise import connections
+
 
 def get_database_config(
 	db_min_connections, db_max_connections,
@@ -84,3 +88,22 @@ def get_tortoise_config(
         "use_tz": False,
         "timezone": "UTC",
     }
+
+# Async context manager to get a Tortoise connection and register pgvector on demand
+class PgVectorConnection:
+    def __init__(self, alias: str = "default"):
+        self.db = connections.get(alias)
+        self.raw = None  # type: ignore
+    
+    async def __aenter__(self) -> asyncpg.Connection:
+        # Acquire a raw asyncpg.Connection
+        self.raw = await self.db._pool.acquire()
+        
+        # tell asyncpg how to handle pgvector
+        await register_vector(self.raw)
+        
+        return self.raw
+    
+    async def __aexit__(self, exc_type, exc, tb):
+        await self.db._pool.release(self.raw)
+        self.raw = None
